+++ CHANGELOG.md
@@ -1,16 +1,21 @@
 ## [0.1.0] - 2015-11-12
 
 ## [0.1.1] - 2015-11-15
+- Initial release
+
 ### Fixed
 - Updated for latest nightly
 
 ## [0.1.2] - 2015-12-01
+
 ### Fixed
 - Updated for latest nightly
 
 ## [0.2.0] - UNRELEASED
+
 ### Fixed
 - Fixed overeager sequence parsing
+
 ### Added
 - Added support for sequences with zero or one repetitions (`$(sequence)?`)
 - Added support for named sequences